Moved database to own file
Previously the database was exposed from the shortener package.
But as further down the line not only the shortener, but also the
login flow (or similar) will make use of it, it makes sense to move
the initialization and exposing the global variable, out of the
shortener scope and contain it in its own package.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>