Prepare for login
This commit has no login flow as of right now, but prepares the code
structure and introduced middleware in order to realize a private and
public section of the page.

- introduced middleware to check login status / hide private pages
Using the `gin-contrib/sessions` package, the server will move a user
back to the login page if they try to access a page hidden behind it.

- changed handling of template files
Previously the templates were loaded and used manually. With e.g.
`gin.LoadTemplatesGlob` this is not necessary anymore. Additionally
handlers become a bit more cleanly to write.

- moved page handlers into their own project

- removed `dashboard.go`
all functionality of this module has been refactored in the new code.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>