Allow configuration using yaml and env
The program is now able to read in a configuration file or be configured
using env variables. This is done using viper

A possible configuration file can be found with `config.example.yaml`.
The available attributes are derived from the `AppConfig` struct in
`config.go`

Note: This commit also includes a `BindStruct` method, which is a hack to
circumvent a few problems with viper trying to `unmarshal` environment
variables. Specifically viper is not able to unmarshal environment
variables for keys in the struct which haven't been mentioned before.
This helper function solves this problem by calling `BindEnv` on each
attribute of the struct (also able to do so with nested structs)

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>