Moved common static files to CDNs
It's generally seen as bad practice to push your own version of e.g.
jQuery or Bootstrap, instead of using a CDN. The latter allows the
browser to cache common documents across domains, a mechanism that
breaks if one serves their own version, even if the document is
identical.

Integrity of the files is maintained, by providing the sha256 for the
linked file so that a compromised CDN, will not become a security issue.

I've chosen jsdelivr.com as the CDN here as it isn't associated with
Cloudflare and is next to cdnjs one of most commonly used ones.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>