Add ability to migrate database on startup
In order to not loose everything and make future updates a huge pain,
this commit introduces database migrations.
Migrations are stored in `server/database/migrations` as sql files and
are applied by the new `RunMigrations` function everytime the database
gets initialized.

Current migrations include the creation of the url database and an INDEX
using the domain and path to speed up runtime access.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>