Add Admin page and URL edit / insert modal
This patch introduces an admin page, that allows managing of all
shortened URLs. They are listed in table with each row having their own
edit and delete button.

As of right now the page can only display information, but not
manipulate them. As such the URL edit or insert modal fills itself with
the corresponding data if editing and has a nice UI, but the buttons
don't actually manipulate the server data.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>