Fix: Shortner now works with subdomains
After changing the underlying data model recently by switching from
storing the full domain to only storing the subdomain, the change wasn't
applied to the actually shortener URL retrieval and redirection code.

With this patch, this is fixed now. Before accessing the database the
last mention of the configured hostname (and potential port numbers) are
removed from the request Host to form the subdomain.

This patch also lower cases the path and subdomain of all requests
before access and when inserting, as URLs presumably should be case
insensitive.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>