fix: URL modal was loaded multiple times
With this patch the URL modal gets only loaded once by the footer if the
current session is done by a user. This fixes a previous issue where in
mobile layout when the navbar disappeared, the modal would only show,
when the hamburger menu is opened.

This also reduces code duplication during page generation.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>