fix: allow empty dashboard subdomain
Previously if one tried to set the primary domain to be the dashboard,
by letting the dashboard subdomain be empty, the server would not make
it possible it access it.

This fix now employs `common.GetHostname` in the main (previously was
generated on its own) and fixes the above mentioned issue.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>