Unify subdomain and path preparation
Previously every query and write to the database needed to sanitize the
subdomain and path on its own. This was relevant for e.g. case
insensitivity, but became more troublesome with the need to remove the
unicode ZWJ.

As such this sanitization is now bundled in one method, that can be
executed every time the database is accessed.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>