feature: Add public attribute to links
Some links should be listed publicly, as they are useful shortcuts to
public resources. For example `g.link.sh -> git.fictional.com`.
Of course this isn't true for every shortened link, so we make use of
a new attribute called `is-public` in the database with this patch.

The attribute is already being listed in the admin overview and can be
changed using the add/edit modal.

Currently the 'is-public' property has no further effect.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>