Add Method that construct origin address
Building the origin address can be a bit tedious as depending on the
presence of a subdomain or path, the format changes. More specifically
a point should occur after the subdomain and a slash before the path.

As this is used multiple times in the templating, having a specific
method for this job makes sense.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>