feature: Add overview page of public URLs
This patch introduces an overview page (`/overview`) that lists all
public URLs with their origin address, target and comment.
In order to make the overview responsive for mobile devices, we use a
grid layout here instead of a table. On larger devices the look should
be similar to a table, but on mobile clients the elements wrap to build
a much nicer overview that doesn't require horizontal scrolling.

The list is build according to the `is_public` property in the database.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>